【back】schema/media.pyをパッケージ化しinput/outputに分離
